Pull parsers out of storeProductConversions and into their own classes
