+++ feature/amazon/src/main/java/com/revenuecat/purchases/amazon/storeProductConversions.kt
@@ -5,12 +5,11 @@ import com.amazon.device.iap.model.Product
 import com.revenuecat.purchases.ProductType
 import com.revenuecat.purchases.common.LogIntent
 import com.revenuecat.purchases.common.MICROS_MULTIPLIER
-import com.revenuecat.purchases.common.OfferingParser
 import com.revenuecat.purchases.common.log
-import com.revenuecat.purchases.models.PurchasingData
-import com.revenuecat.purchases.models.SubscriptionOption
-import com.revenuecat.purchases.models.StoreProduct
 import com.revenuecat.purchases.models.Period
+import com.revenuecat.purchases.models.PurchasingData
+import com.revenuecat.purchases.models.StoreProduct
+import com.revenuecat.purchases.models.SubscriptionOption
 import com.revenuecat.purchases.utils.JSONObjectParceler
 import kotlinx.parcelize.Parcelize
 import kotlinx.parcelize.TypeParceler
@@ -132,11 +131,4 @@ internal fun String.parsePriceUsingRegex(): BigDecimal? {
         price = price.trim()
         BigDecimal(price)
     }
-}
-
-class AmazonOfferingParser : OfferingParser() {
-    override fun Map<String, List<StoreProduct>>.findMatchingProduct(
-        productIdentifier: String,
-        planIdentifier: String?
-    ): StoreProduct? = this[productIdentifier]?.firstOrNull()
 }
