+++ feature/google/src/main/java/com/revenuecat/purchases/google/storeProductConversions.kt
@@ -3,7 +3,6 @@ package com.revenuecat.purchases.google
 import com.android.billingclient.api.ProductDetails
 import com.revenuecat.purchases.ProductType
 import com.revenuecat.purchases.common.LogIntent
-import com.revenuecat.purchases.common.OfferingParser
 import com.revenuecat.purchases.common.log
 import com.revenuecat.purchases.models.GoogleStoreProduct
 import com.revenuecat.purchases.models.Price
@@ -76,24 +75,4 @@ fun List<ProductDetails>.toStoreProducts(): List<StoreProduct> {
         )
     }
     return storeProducts
-}
-
-class GoogleOfferingParser : OfferingParser() {
-    override fun Map<String, List<StoreProduct>>.findMatchingProduct(
-        productIdentifier: String,
-        planIdentifier: String?
-    ): StoreProduct? {
-        if (planIdentifier == null) {
-            // It could be an INAPP or a mis-configured subscription
-            // Try to find INAPP, otherwise null
-            return this[productIdentifier]
-                .takeIf { it?.size == 1 }
-                ?.takeIf { it[0].type == ProductType.INAPP }
-                ?.firstOrNull()
-        }
-        val storeProducts: List<StoreProduct>? = this[productIdentifier]
-        return storeProducts?.firstOrNull { storeProduct ->
-            storeProduct.subscriptionOptions.firstOrNull { it.isBasePlan }?.id == planIdentifier
-        }
-    }
 }
